+++ src/main/java/Commands/DeleteTaskCommand.java
@@ -1,17 +1,10 @@
-//public class DeleteTaskCommand extends Command {
-//    private final int taskNumber;
-//
-//    public DeleteTaskCommand(int taskNumber) {
-//        this.taskNumber = taskNumber;
-//    }
-//
-//    @Override
-//    public void execute(TaskList taskList, Ui ui, Storage storage) {
-//        taskList.deleteTask(taskNumber);
-//        System.out.println("Noted. I've removed this task.");
-//        storage.saveTasks(taskList);
-//    }
-//}
+package Commands;
+
+import TaskLists.TaskList;
+import UI.Ui;
+import Storage.Storage;
+import Tasks.Task;
+
 
 
 public class DeleteTaskCommand extends Command {
