+++ src/main/java/Commands/ExitCommand.java
@@ -1,3 +1,9 @@
+package Commands;
+
+import UI.Ui;
+import TaskLists.TaskList;
+import Storage.Storage;
+
 public class ExitCommand extends Command {
 
     @Override
