+++ src/main/java/UI/Ui.java
@@ -1,3 +1,5 @@
+package UI;
+
 import java.util.Scanner;
 
 public class Ui {
