+++ src/main/java/data/Jack.txt
@@ -12,3 +12,4 @@ T | 1 | run
 T | 0 | run 
 E | 0 | project meeting | Mon 2pm to 4pm
 D | 1 | run | 2003-10-12
+D | 1 |  | 2003-10-12
